requirements.txt
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0
fastapi>=0.100.0
uvicorn>=0.20.0
scikit-learn>=1.0.0
numpy>=1.21.0
pydantic>=2.0.0
torch>=2.0.0
